move reports and rubocop
