+++ app/controllers/reports/status_filter_params.rb
@@ -1,5 +1,7 @@
 module Reports
+
   module StatusFilterParams
+
     extend ActiveSupport::Concern
 
     included do
@@ -28,5 +30,7 @@ module Reports
 
       @date_range_field = params[:date_range_field] || "journal_or_statement_date"
     end
+
   end
+
 end
