+++ spec/controllers/reports/instrument_day_reports_controller_spec.rb
