+++ spec/controllers/reports/instrument_reports_controller_spec.rb
